fixitpro_light/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
pyodbc==5.0.1
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
eventlet==0.35.1
